% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReferenceProjection.R
\name{createReferenceProjection}
\alias{createReferenceProjection}
\title{Create projection matrix and reference profiles from a beta matrix of reference samples.}
\usage{
createReferenceProjection(
  beta.ref,
  cell.type,
  n.top.probe = 150,
  top.probes = NULL
)
}
\arguments{
\item{beta.ref}{A matrix of beta values. Probes are in rows, samples in columns. Row names have to be probe IDs.}

\item{cell.type}{A vector of cell names. Should have the same length as columns of beta.ref.}

\item{n.top.probe}{Number of probes for each cell type to be included for the projection.}

\item{top.probes}{A vector of probe IDs. If provided, no probe selection will be performed and the projection will be created based on these probes.}
}
\value{
A list with 3 elements:
\itemize{
\item "projection" is the projection matrix to convert selected probes to PCA space.Row names have to be probe IDs. Columns are PCs;
\item "ref.beta.mat" is the matrix of mean beta values for the reference cells. Rows are probes, columns samples. Row names have to be probe IDs;
\item "pca_res" is the PCA result from prcomp.
}
}
\description{
Create projection matrix and reference profiles from a beta matrix of reference samples.
}
\examples{
\dontrun{
ref.projection.EPIC = createReferenceProjection(beta.ref=ref.beta, cell.type = ref.cell,n.top.probe = 150)
}
}
