% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellComposition.R
\name{estimateCellComposition}
\alias{estimateCellComposition}
\title{Estimate cell compostions in test samples based on reference profile and projection.}
\usage{
estimateCellComposition(
  test.beta,
  ref.beta.mat,
  projection,
  n.PC = 20,
  extended = TRUE
)
}
\arguments{
\item{test.beta}{Beta matrix of samples to be tested. Probes are in rows. Samples are in columns. Row names have to be probe IDs.}

\item{ref.beta.mat}{Matrix of mean beta values for the reference cells. Row names have to be probe IDs. For EPIC arrays,
ref.projection.EPIC$ref.beta.mat, pre-calculated from GSE167998, can be used.}

\item{projection}{Projection matrix to convert selected probes to PCA space. Row names have to be probe IDs. Columns are PCs.For EPIC arrays,
ref.projection.EPIC$projection, pre-calculated from GSE167998, can be used.}

\item{n.PC}{number of PCs to include for deconvolution.}

\item{extended}{Boolean variable to indicate whether to include Bcell, CD4T, CD8T and Gran in output.}
}
\value{
A matrix of cell compositions in the test samples. Samples are in rows, cell types in columns.
}
\description{
Estimate cell compostions in test samples based on reference profile and projection.
}
\examples{
\dontrun{
test.res=estimateCellComposition(test.beta = test.beta, ref.beta.mat = ref.projection.EPIC$ref.beta.mat,projection = ref.projection.EPIC$projection, n.PC = 20,extended = F)
plot_celltype(test.res*100,test.pd*100,celltype = 'Bmem')
}
}
