% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_celltype.R
\name{plot_celltype}
\alias{plot_celltype}
\title{Plot predicted against actual cell compositions for specific cell type.}
\usage{
plot_celltype(predicted, actual, celltype, plot = T)
}
\arguments{
\item{predicted}{A data.frame of predicted cell compositions. Samples are in rows. Cell types are in columns.}

\item{actual}{A data.frame of actual cell compositions. Samples are in rows. Cell types are in columns.}

\item{celltype}{A string variable containing a single cell type to plot. Must be present as a column name in both predicted and actual data.frames.}

\item{plot}{A boolean variable to indicate whether to plot.}
}
\value{
A list with 2 elements:
\itemize{
\item A data.frame of 3 columns: Predicted, Actual, and celltype.
\item A ggplot of the scatter plot of predicted against actual values for the cell type.
}
}
\description{
Plot predicted against actual cell compositions for specific cell type.
}
\examples{
\dontrun{
test.res=estimateCellComposition(test.beta = test.beta, ref.beta.mat = ref.projection.EPIC$ref.beta.mat,projection = ref.projection.EPIC$projection, n.PC = 20,extended = F)
plot_celltype(test.res*100,test.pd*100,celltype = 'Bmem')
}
}
